new class - AnimalRequirements
