+++ RUFAS/routines/animal/life_cycle/heiferII.py
@@ -19,7 +19,7 @@ from RUFAS.routines.animal.life_cycle.heiferI import HeiferI
 from RUFAS.routines.animal.life_cycle.animal_base import AnimalBase
 from RUFAS.routines.animal.manure.growing_heifer_manure_excretion import \
     manure_calculations
-from RUFAS.routines.animal.ration.animal_requirements import calc_rqmts
+from RUFAS.routines.animal.ration.animal_requirements import AnimalRequirements
 from random import random
 import math
 from RUFAS.routines.animal.life_cycle import animal_constants as const
@@ -211,21 +211,22 @@ class HeiferII(HeiferI):
         """
 		Calculates this heiferII's nutrient requirements.
 		"""
-        req = calc_rqmts(body_weight=self.body_weight,
+        req = AnimalRequirements()
+        animal_requirements = req.calc_rqmts(body_weight=self.body_weight,
                          mature_body_weight=self.mature_body_weight,
                          day_of_pregnancy=self.days_in_preg,
                          animal_type=animal_grouping_scenario.get_animal_type(self),
                          body_condition_score_5 = 3,
                          previous_temperature = temp,
                          average_daily_gain_heifer = self.daily_growth)
-        self.NEmaint = req['NEmaint']
-        self.NEg = req['NEg']
-        self.NEpreg = req['NEpreg']
-        self.NEl = req['NEl']
-        self.MP_req = req['MP_req']
-        self.Ca_req = req['Ca_req']
-        self.P_req = req['P_req']
-        self.DMIest = req['DMIest']
+        self.NEmaint = animal_requirements['NEmaint']
+        self.NEg = animal_requirements['NEg']
+        self.NEpreg = animal_requirements['NEpreg']
+        self.NEl = animal_requirements['NEl']
+        self.MP_req = animal_requirements['MP_req']
+        self.Ca_req = animal_requirements['Ca_req']
+        self.P_req = animal_requirements['P_req']
+        self.DMIest = animal_requirements['DMIest']
 
     def calc_manure_excretion(self, feed, methane_model):
         """
